Clean up functions and reorganize file structure
