+++ tools/__init__.py
@@ -1,3 +1,4 @@
+from .checks import manage_messages, manage_channels
 from .decorators import check_general, check_event_id, check_title, check_url, check_date_time
 from .exceptions import GuildNotRegistered, GuildNotEnabled, MessageUnreachable, BadInput
 from .parsers import parse_time, parse_date, parse_type
